Add update method in GoogleSheetsAuctionRepository and move creating row (List) from auction to separate file - CreateRowInGoogleSheets.
