+++ backend/src/main/java/com/example/backend/repository/AuctionRepository.java
@@ -10,6 +10,7 @@ import java.util.UUID;
 @Repository
 public interface AuctionRepository {
     void save(Auction auction) throws IOException;
+    void update(UUID auctionId, Auction auction) throws IOException;
     List<Auction> findAllByFiltersAndUser(List<String> statuses, Boolean myAuctions, Boolean followed, List<String> dates, String userEmail) throws IOException;
     void follow(UUID auctionId, String userEmail) throws IOException;
     void unfollow(UUID auctionId, String userEmail) throws IOException;
