+++ backend/src/main/java/com/example/backend/repository/GoogleSheetsAuctionRepository.java
@@ -3,9 +3,11 @@ package com.example.backend.repository;
 import com.example.backend.model.Auction;
 import com.example.backend.util.AuctionQuery;
 import com.example.backend.service.GoogleSheetsService;
+import com.example.backend.util.CreateRowInGoogleSheets;
 import com.example.backend.util.FollowersQuery;
 import com.example.backend.util.GvizResponseParser;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import java.io.IOException;
@@ -18,40 +20,38 @@ public class GoogleSheetsAuctionRepository implements AuctionRepository {
     private final GoogleSheetsService googleSheetsService;
     private final GvizResponseParser gvizResponseParser;
 
+    @Autowired
+    private CreateRowInGoogleSheets createRowInGoogleSheets;
+
     public GoogleSheetsAuctionRepository(GoogleSheetsService googleSheetsService, GvizResponseParser gvizResponseParser) {
         this.googleSheetsService = googleSheetsService;
         this.gvizResponseParser = gvizResponseParser;
     }
 
-    private String makeNotNull(Object value) {
-        return value != null ? value.toString() : "";
-    }
-    private Object makeNotNullForNumbers(Object value) {
-        return value != null ? value : "";
+    @Override
+    public void save(Auction auction) throws IOException {
+        List<Object> row = createRowInGoogleSheets.createRow(auction);
+
+        googleSheetsService.appendRow("Auction", List.of(row));
     }
 
     @Override
-    public void save(Auction auction) throws IOException {
-        List<Object> row = Arrays.asList(
-                makeNotNull(auction.getId()),
-                makeNotNull(auction.getModeratorEmail()),
-                makeNotNull(auction.getAuctionDate()),
-                makeNotNull(auction.getAuctionDate()),
-                makeNotNull(auction.getSupplierName()),
-                makeNotNull(auction.getSupplierEmail()),
-                makeNotNull(auction.getTitle()),
-                makeNotNull(auction.getDescription()),
-                makeNotNull(auction.getFileId()),
-                makeNotNull(auction.getCity()),
-                makeNotNullForNumbers(auction.getStartingPrice()),
-                makeNotNull(auction.getFollowers()),
-                makeNotNullForNumbers(auction.getFollowersCount()),
-                makeNotNull(auction.getSlackThreadLink()),
-                makeNotNullForNumbers(auction.getCurrentBid()),
-                makeNotNull(auction.getWinner())
-        );
+    public void update(UUID auctionId, Auction auction) throws IOException {
+        List<Object> updatedRow = createRowInGoogleSheets.createRow(auction);
 
-        googleSheetsService.appendRow("Auction", List.of(row));
+        try {
+            List<List<Object>> rows = googleSheetsService.readSheet("Auction");
+            for (int i = 0; i < rows.size(); i++) {
+                List<Object> searchRow = rows.get(i);
+                if (!searchRow.isEmpty() && auctionId.toString().equals(searchRow.get(0).toString())) {
+                    googleSheetsService.updateRow("Auction", i, updatedRow);
+                    return;
+                }
+            }
+            throw new RuntimeException("Auction not found: " + auctionId);
+        } catch (IOException e) {
+            throw new RuntimeException("Failed to update auction", e);
+        }
     }
 
     @Override
@@ -79,7 +79,7 @@ public class GoogleSheetsAuctionRepository implements AuctionRepository {
                     // TODO: Replace hardcoded column index 11 with dynamic detection of the "followers" column
                     googleSheetsService.updateCellValue("Auction", i, 11, json);
                     // This should probably be replaced with a single batched request to reduce latency and API usage
-                    googleSheetsService.updateCellValue("Auction", i, 12, makeNotNull(followers.size()));
+                    googleSheetsService.updateCellValue("Auction", i, 12, createRowInGoogleSheets.makeNotNull(followers.size()));
                     return;
                 }
             }
